backend/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=2.1.1
numpy>=1.24.3
scikit-learn>=1.3.1
flask>=3.0.0
flask-cors>=4.0.0
PyMuPDF>=1.23.5
Pillow>=10.0.1
opencv-python>=4.8.1.78
plotly>=5.18.0
tqdm>=4.66.1
scipy>=1.11.3
pdfplumber>=0.10.3
pytesseract>=0.3.10
pyarrow>=13.0.0
joblib>=1.3.2
werkzeug>=3.0.1
datasets>=2.14.5
huggingface-hub>=0.19.4